+++ packages/client/src/StreamFactory.ts
@@ -8,6 +8,7 @@ import { StreamStorageRegistry } from './registry/StreamStorageRegistry'
 import { Stream, StreamrStreamConstructorOptions } from './Stream'
 import { Resends } from './subscribe/Resends'
 import { Subscriber } from './subscribe/Subscriber'
+import { LoggerFactory } from './utils/LoggerFactory'
 
 @scoped(Lifecycle.ContainerScoped)
 export class StreamFactory {
@@ -18,6 +19,7 @@ export class StreamFactory {
     private readonly streamRegistryCached: StreamRegistryCached
     private readonly streamRegistry: StreamRegistry
     private readonly streamStorageRegistry: StreamStorageRegistry
+    private readonly loggerFactory: LoggerFactory
     private readonly eventEmitter: StreamrClientEventEmitter
     private readonly timeoutsConfig: TimeoutsConfig
 
@@ -28,6 +30,7 @@ export class StreamFactory {
         @inject(delay(() => StreamRegistryCached)) streamRegistryCached: StreamRegistryCached,
         @inject(delay(() => StreamRegistry)) streamRegistry: StreamRegistry,
         streamStorageRegistry: StreamStorageRegistry,
+        loggerFactory: LoggerFactory,
         eventEmitter: StreamrClientEventEmitter,
         @inject(ConfigInjectionToken.Timeouts) timeoutsConfig: TimeoutsConfig
     ) {
@@ -37,6 +40,7 @@ export class StreamFactory {
         this.streamRegistryCached = streamRegistryCached
         this.streamRegistry = streamRegistry
         this.streamStorageRegistry = streamStorageRegistry
+        this.loggerFactory = loggerFactory
         this.eventEmitter = eventEmitter
         this.timeoutsConfig = timeoutsConfig
     }
@@ -50,6 +54,7 @@ export class StreamFactory {
             this.streamRegistryCached,
             this.streamRegistry,
             this.streamStorageRegistry,
+            this.loggerFactory,
             this.eventEmitter,
             this.timeoutsConfig
         )